requirements.txt
Flask==1.1.2
Pillow==8.1.0
Werkzeug==1.0.1
